% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{c_area}
\alias{c_area}
\title{Area of a circle}
\usage{
c_area(radius)
}
\arguments{
\item{radius}{The radius of the circle.}
}
\value{
The area of the circle.
}
\description{
This function calculates the area of a circle given its radius.
}
\examples{
c_area(5)
}
