% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{c_circumference}
\alias{c_circumference}
\title{Circumference of a circle}
\usage{
c_circumference(radius)
}
\arguments{
\item{radius}{The radius of the circle.}
}
\value{
The circumference of the circle.
}
\description{
This function calculates the circumference of a circle given its radius.
}
\examples{
c_circumference(5)
}
